fix: add interface OrderService and implement it
Took 14 seconds
